cloud_functions/gcs_to_bq/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.11.0
